Create README.md
I have intentionally selected a subject of great significance. The Covid-19 pandemic has undeniably left an indelible mark on the world, ushering in an era of unprecedented changes. This project serves as an opportunity for me to delve deeper into the data, seeking to gain valuable insights into the global impact of COVID-19, particularly about the reported cases and fatalities.